move metrics to its own file
