+++ src/backends/plonky2/circuits/mod.rs
@@ -1,4 +1,5 @@
 pub mod common;
 pub mod mainpod;
+pub mod metrics;
 pub mod signedpod;
 pub mod utils;
